Small errors and getting rating to work

+++ Backend/Player.h
@@ -1,75 +1,3 @@
-// #pragma once
-
-// #include <string>
-// #include <vector>
-// #include "Inventory.h"
-// #include "Greenhouse.h"
-// #include "Worker.h"
-// #include "Memento.h"
-
-// // Frontend
-// class InventoryUI;
-
-// class Player
-// {
-
-//     InventoryUI *inventoryUI;
-
-// public:
-//     Player();
-//     ~Player();
-    
-//     float getMoney() const;
-//     void setMoney(float amount);
-//     void addMoney(float amount);
-//     void subtractMoney(float amount);
-//     void UpdateGameTime(float dt);
-    
-//     int getRating() const;
-//     void setRating(int r);
-//     void addRating(int r);
-//     void subtractRating(int r);
-    
-//     int getDay() const;
-//     int getHour() const;
-//     int getMinute() const;
-//     void setTime(int d, int h, int m);
-//     void advanceTime(int minutes);
-    
-//     Inventory* getInventory() const;
-//     Greenhouse* getPlot() const;
-    
-//     void addWorker(Worker* worker);
-//     void fireWorker(int index);
-//     Worker* getWorker(int index) const;
-//     int getWorkerCount() const;
-//     const std::vector<Worker*>& getWorkers() ;
-
-    
-//     void setProtected(bool prot);
-//     bool isProtected();
-
-//     Memento* createMemento() const;
-//     void setMemento(Memento* memento);
-//     std::string getTimeString() const;
-//     InventoryUI *getInventoryUI() const { return inventoryUI; }
-
-    
-//     private:
-//     float money;
-//     int rating;
-//     int day;
-//     int hour;
-//     int minute;
-//     bool safe;
-
-//     float timeAccumulator = 0.0f;
-    
-//     Inventory* inventory;
-//     Greenhouse* plot;
-//     std::vector<Worker*> workers;
-// };
-
 #pragma once
 
 #include <string>
@@ -91,10 +19,10 @@ public:
     void addMoney(float amount);
     void subtractMoney(float amount);
     
-    int getRating() const;
-    void setRating(int r);
-    void addRating(int r);
-    void subtractRating(int r);
+    float getRating() const;
+    void setRating(float r);
+    void addRating(float r);
+    void subtractRating(float r);
     
     int getDay() const;
     int getHour() const;
@@ -125,7 +53,7 @@ public:
 
 private:
     float money;
-    int rating;
+    float rating;
     int day;
     int hour;
     int minute;
